Add libvts (Virtual Terminal Session) api
This adds a small library - libvts - to the cogl repo for safely
handling the privileged operations involved in opening a tty, opening a
pam session (optional), claiming drmMaster() mode for using KMS and
opening evdev input devices for the session-seat.

This was written to support the launching of Wayland based compositors
but there is nothing Wayland specific about the code so it can be used
to launch any kind of full screen application that wants to run on a
Linux virtual console via KMS.

The code has very minimal dependencies (no glib usage for example) and
could be factored out into a standalone project. This would enable
weston-launch to share the same code.

In the meantime we will maintain this code in Cogl so that applications
using the KMS backend can automatically take advantage of it.