Add Android.mk based build support
This makes it possible to build Cogl for Android using ndk-build instead
of autotools. This is the typical way to build software for Android and
much simpler than using autotools given that there's no need to
dynamically probe the platform being built for. To make it easy to
maintain this way of building Cogl this moves the list of source files
that need to be built into a Makefile.sources file that can be included
by cogl/Makefile.am and Android.mk